requirements.txt
openclaw>=2026.3.0
requests>=2.31.0
pyyaml>=6.0.1
schedule>=1.2.0
python-dateutil>=2.8.2
beautifulsoup4>=4.12.0
feedparser>=6.0.10
jinja2>=3.1.2
pillow>=10.0.0
aiohttp>=3.9.0
asyncio>=3.4.3
flask>=2.3.0
openai>=1.0.0
playwright>=1.40.0

# RPA 浏览器自动化
playwright>=1.40.0
pyperclip>=1.8.2
pyautogui>=0.9.54
